Make sendThreaded return an event response
This lets us use sendThreaded instead of direct client interactions to
send messages with thread data from a caller, which allows going through
the same Markdown handling and generally through the adapter shim and
any additional enrichment it is planning to do.

To do this, the sendURL helper had to be adjusted to properly return a
Promise instead of a request event.