Correctly default markdown to true
The previous version failed to interpret markdown if the passed envelope
did not carry a message with a metadata property, i.e. any usage that
wasn't a custom MatrixMessage. The logic now correctly defaults to true
in these cases, so only MatrixMessages with interpretMarkdown explicitly
set to false skip Markdown interpretation.