Incoming room data uses canonical alias as allowed
Incoming message envelopes attempt to resolve the canonical alias and
use it, so that any additional storage or lookup of room information can
present the user with what is ostensibly the "best" presentation of a
user-friendly name. Since the internals of message sending already
support sending to a room by alias, this should not affect
deliverability or other issues.